+++ features/encrypt/core/src/main/java/org/apache/shardingsphere/encrypt/rewrite/token/pojo/EncryptFunctionAssignmentToken.java
@@ -18,6 +18,7 @@
 package org.apache.shardingsphere.encrypt.rewrite.token.pojo;
 
 import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.infra.database.core.metadata.database.enums.QuoteCharacter;
 
 import java.util.Collection;
 import java.util.LinkedList;
@@ -31,8 +32,8 @@ public final class EncryptFunctionAssignmentToken extends EncryptAssignmentToken
     
     private final Collection<FunctionAssignment> assignments = new LinkedList<>();
     
-    public EncryptFunctionAssignmentToken(final int startIndex, final int stopIndex) {
-        super(startIndex, stopIndex);
+    public EncryptFunctionAssignmentToken(final int startIndex, final int stopIndex, final QuoteCharacter quoteCharacter) {
+        super(startIndex, stopIndex, quoteCharacter);
     }
     
     /**
@@ -42,7 +43,7 @@ public final class EncryptFunctionAssignmentToken extends EncryptAssignmentToken
      * @param value assignment value
      */
     public void addAssignment(final String columnName, final Object value) {
-        FunctionAssignment functionAssignment = new FunctionAssignment(columnName, value);
+        FunctionAssignment functionAssignment = new FunctionAssignment(columnName, value, getQuoteCharacter());
         assignments.add(functionAssignment);
         builder.append(functionAssignment).append(", ");
     }
@@ -68,9 +69,11 @@ public final class EncryptFunctionAssignmentToken extends EncryptAssignmentToken
         
         private final Object value;
         
+        private final QuoteCharacter quoteCharacter;
+        
         @Override
         public String toString() {
-            return String.format("%s = %s", columnName, value);
+            return quoteCharacter.wrap(columnName) + " = " + value;
         }
     }
 }