+++ kernel/metadata/core/src/main/java/org/apache/shardingsphere/metadata/persist/node/ReservationNodePath.java
@@ -21,31 +21,22 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 /**
- * Process node.
+ * Reservation node path.
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class ProcessNode {
+public final class ReservationNodePath {
     
-    private static final String EXECUTION_NODES = "execution_nodes";
+    private static final String ROOT_NODE = "reservation";
     
-    /**
-     * Get process id path.
-     *
-     * @param processId process id
-     * @return execution path
-     */
-    public static String getProcessIdPath(final String processId) {
-        return String.join("/", "", EXECUTION_NODES, processId);
-    }
+    private static final String WORKER_ID_NODE = "worker_id";
     
     /**
-     * Get process list instance path.
+     * Get worker id reservation path.
      *
-     * @param processId process id
-     * @param instancePath instance path
-     * @return execution path
+     * @param workerId worker id
+     * @return worker id reservation path
      */
-    public static String getProcessListInstancePath(final String processId, final String instancePath) {
-        return String.join("/", "", EXECUTION_NODES, processId, instancePath);
+    public static String getWorkerIdReservationPath(final int workerId) {
+        return String.join("/", "", ROOT_NODE, WORKER_ID_NODE, String.valueOf(workerId));
     }
 }