+++ mode/core/src/main/java/org/apache/shardingsphere/mode/persist/service/unified/StatePersistService.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.mode.persist.service.unified;
 import com.google.common.base.Strings;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.mode.state.ClusterState;
-import org.apache.shardingsphere.metadata.persist.node.StatesNode;
+import org.apache.shardingsphere.metadata.persist.node.StatesNodePath;
 import org.apache.shardingsphere.mode.spi.PersistRepository;
 
 /**
@@ -37,7 +37,7 @@ public final class StatePersistService {
      * @param state to be updated cluster state
      */
     public void update(final ClusterState state) {
-        repository.persist(StatesNode.getClusterStateNodePath(), state.name());
+        repository.persist(StatesNodePath.getClusterStatePath(), state.name());
     }
     
     /**
@@ -46,7 +46,7 @@ public final class StatePersistService {
      * @return loaded cluster state
      */
     public ClusterState load() {
-        String value = repository.query(StatesNode.getClusterStateNodePath());
+        String value = repository.query(StatesNodePath.getClusterStatePath());
         return Strings.isNullOrEmpty(value) ? ClusterState.OK : ClusterState.valueOf(value);
     }
 }