+++ test/e2e/operation/transaction/src/test/java/org/apache/shardingsphere/test/e2e/transaction/cases/alterresource/SwitchingTransactionRuleTestCase.java
@@ -107,12 +107,12 @@ public final class SwitchingTransactionRuleTestCase extends BaseTransactionTestC
                 alterLocalTransactionRule();
                 Awaitility.await().atMost(20L, TimeUnit.SECONDS).pollInterval(19L, TimeUnit.SECONDS).until(() -> true);
                 alterXaTransactionRule("Narayana");
+                Awaitility.await().atMost(20L, TimeUnit.SECONDS).pollInterval(19L, TimeUnit.SECONDS).until(() -> true);
                 if (SWITCH_COUNT.incrementAndGet() >= MAX_SWITCH_COUNT) {
                     alterLocalTransactionRule();
                     IS_FINISHED.set(true);
                     break;
                 }
-                Awaitility.await().atMost(20L, TimeUnit.SECONDS).pollInterval(19L, TimeUnit.SECONDS).until(() -> true);
             }
         }